backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# ORM & Database
sqlmodel==0.0.22
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
pyjwt[crypto]==2.10.1
httpx==0.28.1

# Storage (Cloudflare R2 / S3-compatible)
boto3==1.36.4

# Payments
stripe==11.4.1

# Utilities
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
